backend/requirements.txt
# xFinance Backend - FastAPI
# Python 3.10+

# Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Autenticação
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# CORS e segurança
python-multipart>=0.0.6

# Banco de dados
# SQLite já incluso no Python

# Utilitários
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

